Aquisicao de dados remota (via Retrofit) implementadas - implementacao basica -. Refac em diversas classes.
