--- app/src/main/java/com/seletiva/santander/investment/Models/events/SendButtonClickEvent.java
@@ -1,7 +1,0 @@
-package com.seletiva.santander.investment.Models.events;
-
-public class SendButtonClickEvent {
-    public static SendButtonClickEvent newClickEvent() {
-        return new SendButtonClickEvent();
-    }
-}
