--- app/src/main/java/com/seletiva/santander/investment/UI/Activities/BaseActivity.java
@@ -1,8 +1,0 @@
-package com.seletiva.santander.investment.UI.Activities;
-
-import android.support.v7.app.AppCompatActivity;
-
-public class BaseActivity extends AppCompatActivity {
-
-}
-
