+++ app/src/main/java/com/seletiva/santander/investment/controllers/BaseResponse.java
@@ -1,0 +1,6 @@
+package com.seletiva.santander.investment.controllers;
+
+public interface BaseResponse {
+    void onConnectionFailure();
+    void onRequestFailure(int stringMessageId);
+}
