+++ app/src/main/java/com/seletiva/santander/investment/controllers/cells/CellsDataSource.java
@@ -1,0 +1,11 @@
+package com.seletiva.santander.investment.controllers.cells;
+
+import com.seletiva.santander.investment.ui.form.domain.model.CellHolder;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+
+public interface CellsDataSource {
+    @GET("cells.json")
+    Call<CellHolder> getCells();
+}
