+++ app/src/main/java/com/seletiva/santander/investment/controllers/funds/FundsDataSource.java
@@ -1,0 +1,9 @@
+package com.seletiva.santander.investment.controllers.funds;
+
+import retrofit2.Call;
+import retrofit2.http.GET;
+
+public interface FundsDataSource {
+    @GET("funds.json")
+    Call<Object> getFunds();
+}
