+++ app/src/main/java/com/seletiva/santander/investment/controllers/funds/FundsResponse.java
@@ -1,0 +1,7 @@
+package com.seletiva.santander.investment.controllers.funds;
+
+import com.seletiva.santander.investment.controllers.BaseResponse;
+
+public interface FundsResponse extends BaseResponse {
+    void onFundsResponse(Object cellHolder);
+}
