+++ app/src/main/java/com/seletiva/santander/investment/services/ApiClient.java
@@ -1,0 +1,20 @@
+package com.seletiva.santander.investment.services;
+
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
+
+public class ApiClient {
+    private static Retrofit retrofit;
+    //FIXME: convert to build time variable
+    private static final String BASE_URL = "https://floating-mountain-50292.herokuapp.com/";
+
+    public static Retrofit getInstance() {
+        if (retrofit == null) {
+            retrofit = new retrofit2.Retrofit.Builder()
+                    .baseUrl(BASE_URL)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build();
+        }
+        return retrofit;
+    }
+}
