+++ app/src/main/java/com/seletiva/santander/investment/ui/BaseActivity.java
@@ -1,0 +1,7 @@
+package com.seletiva.santander.investment.ui;
+
+import android.support.v7.app.AppCompatActivity;
+
+public class BaseActivity extends AppCompatActivity {
+
+}
