+++ app/src/main/java/com/seletiva/santander/investment/ui/BasePresenter.java
@@ -1,0 +1,5 @@
+package com.seletiva.santander.investment.ui;
+
+public interface BasePresenter {
+    void start();
+}
