+++ app/src/main/java/com/seletiva/santander/investment/ui/BaseView.java
@@ -1,0 +1,5 @@
+package com.seletiva.santander.investment.ui;
+
+public interface BaseView<T extends BasePresenter> {
+    void setPresenter(T presenter);
+}
