+++ app/src/main/java/com/seletiva/santander/investment/ui/form/domain/model/CellType.java
@@ -1,0 +1,19 @@
+package com.seletiva.santander.investment.ui.form.domain.model;
+
+public enum CellType {
+    field(1),
+    text(2),
+    image(3),
+    checkbox(4),
+    send(5);
+
+    private int type;
+
+    CellType(int type) {
+        this.type = type;
+    }
+
+    public int getType() {
+        return type;
+    }
+}
