+++ app/src/main/java/com/seletiva/santander/investment/ui/form/domain/model/SendButtonClickEvent.java
@@ -1,0 +1,7 @@
+package com.seletiva.santander.investment.ui.form.domain.model;
+
+public class SendButtonClickEvent {
+    public static SendButtonClickEvent newClickEvent() {
+        return new SendButtonClickEvent();
+    }
+}
