+++ app/src/main/java/com/seletiva/santander/investment/ui/view/FormComponentView.java
@@ -1,0 +1,153 @@
+package com.seletiva.santander.investment.ui.view;
+
+import android.content.Context;
+import android.graphics.PorterDuff;
+import android.support.annotation.Nullable;
+import android.util.AttributeSet;
+import android.view.View;
+import android.widget.Button;
+import android.widget.CheckBox;
+import android.widget.EditText;
+import android.widget.ImageButton;
+import android.widget.LinearLayout;
+import android.widget.TextView;
+
+import com.seletiva.santander.investment.ui.form.domain.model.Cell;
+import com.seletiva.santander.investment.ui.form.domain.model.CellType;
+import com.seletiva.santander.investment.ui.form.domain.model.SendButtonClickEvent;
+import com.seletiva.santander.investment.R;
+import com.seletiva.santander.investment.utils.StringUtils;
+
+import org.greenrobot.eventbus.EventBus;
+
+public class FormComponentView extends LinearLayout implements FormFieldListener {
+    private Cell cellCore;
+    private EditText optionalTextField;
+
+    public FormComponentView(Context context) {
+        super(context);
+    }
+
+    public FormComponentView(Context context, @Nullable AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public FormComponentView(Context context, @Nullable AttributeSet attrs, int defStyleAttr) {
+        super(context, attrs, defStyleAttr);
+    }
+
+    public void inflateWithCellModel(Cell core) {
+        cellCore = core;
+
+        if (core != null) {
+            switch (cellCore.getType()) {
+                case field:
+                    inflate(getContext(), R.layout.form_component_field, this);
+                    configureComponentTitle();
+                    configureClearButton();
+                    configureTextualDataWatcher();
+                    break;
+
+                case text:
+                    inflate(getContext(), R.layout.form_component_text, this);
+                    configureTextualDisplayData();
+                    break;
+
+                case checkbox:
+                    inflate(getContext(), R.layout.form_component_checkbox, this);
+                    configureCheckBoxData();
+                    break;
+
+                case send:
+                    inflate(getContext(), R.layout.form_component_button, this);
+                    configureCallToAction();
+                    break;
+
+                default:
+                    break;
+            }
+        }
+    }
+
+    private void configureCallToAction() {
+        final Button button = findViewById(R.id.callToAction);
+        button.setText(cellCore.getMessage());
+
+        button.setOnClickListener(new OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                EventBus.getDefault().post(SendButtonClickEvent.newClickEvent());
+            }
+        });
+    }
+
+    private void configureCheckBoxData() {
+        final CheckBox checkBox = findViewById(R.id.checkBox);
+        checkBox.setText(cellCore.getMessage());
+    }
+
+    private void configureTextualDisplayData() {
+        final TextView title = findViewById(R.id.textualDisplayData);
+        title.setText(cellCore.getMessage());
+    }
+
+    private void configureComponentTitle() {
+        final TextView title = findViewById(R.id.componentTitle);
+        title.setText(cellCore.getMessage());
+    }
+
+    private void configureTextualDataWatcher() {
+        FormFieldWatcher watcher = new FormFieldWatcher(this);
+
+        switch (cellCore.getTypeField()) {
+            case telNumber:
+                watcher.enablePhoneValidationMode();
+                break;
+            case email:
+                watcher.enableMailValidationMode();
+                break;
+            default:
+                break;
+        }
+
+        optionalTextField = findViewById(R.id.textualInputData);
+        optionalTextField.addTextChangedListener(watcher);
+    }
+
+    private void configureClearButton() {
+        ImageButton clearButton = findViewById(R.id.clearButton);
+        final EditText textualInputData = findViewById(R.id.textualInputData);
+
+        clearButton.setOnClickListener(new OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                textualInputData.setText(null);
+            }
+        });
+    }
+
+    @Override
+    public void updateEditTextColor(int color) {
+        optionalTextField.getBackground().setColorFilter(getResources().getColor(color),
+                PorterDuff.Mode.SRC_ATOP);
+    }
+
+    public boolean isValid() {
+        if (cellCore.getType() == CellType.field) {
+            String inputData = optionalTextField.getText().toString();
+
+            switch (cellCore.getTypeField()) {
+                case telNumber:
+                    return StringUtils.isPhoneNumberValid(inputData);
+
+                case email:
+                    return StringUtils.validateEmailAdress(inputData);
+
+                default:
+                    break;
+            }
+        }
+
+        return true;
+    }
+}
