+++ app/src/main/java/com/seletiva/santander/investment/ui/view/FormFieldListener.java
@@ -1,0 +1,9 @@
+package com.seletiva.santander.investment.ui.view;
+
+public interface FormFieldListener {
+    /**
+     * This method shall be called after the text changes. Than, a new tint color will be requested.
+     * @param color A new color
+     */
+    void updateEditTextColor(int color);
+}
